refactor(check): undo recent check rename
Rename check "no-focusable-content-for-frame" back to "frame-focusable-content".
