+++ lib/checks/keyboard/frame-focusable-content.json
@@ -1,5 +1,5 @@
 {
-  "id": "no-focusable-content-for-frame",
+  "id": "frame-focusable-content",
   "evaluate": "no-focusable-content-for-frame-evaluate",
   "metadata": {
     "impact": "serious",
