+++ lib/core/base/metadata-function-map.js
@@ -96,7 +96,7 @@ import focusableModalOpenEvaluate from '../../checks/keyboard/focusable-modal-op
 import focusableNoNameEvaluate from '../../checks/keyboard/focusable-no-name-evaluate';
 import focusableNotTabbableEvaluate from '../../checks/keyboard/focusable-not-tabbable-evaluate';
 import landmarkIsTopLevelEvaluate from '../../checks/keyboard/landmark-is-top-level-evaluate';
-import noFocusableContentForFrameEvaluate from '../../checks/keyboard/no-focusable-content-for-frame-evaluate';
+import noFocusableContentForFrameEvaluate from '../../checks/keyboard/frame-focusable-content-evaluate';
 import noFocusableContentForAriaTextEvaluate from '../../checks/keyboard/no-focusable-content-for-aria-text-evaluate';
 import noFocusableContentForNestedInteractiveEvaluate from '../../checks/keyboard/no-focusable-content-for-nested-interactive-evaluate';
 import tabindexEvaluate from '../../checks/keyboard/tabindex-evaluate';
